Removed factory methods from 2 classes.
ControllerClass and StartNodeEventHandler had a factory method to create new objects, because some members required "this", and "this" shouldn't escape the constructor.

Turns out that this is really only important in multi-threaded applications, which isn't the case here.

By removing the separate initializers, the members can be made final, making the classes immutable.