Use java.util.Optional inside the implementation where possible.
In the quest to eradicate all kinds of null-checks, the code internally
uses Optionals now.

Because of that there's now an UncheckedXmlProcessorException that
wraps a checked exception in an unchecked one, so that lamdba's can
be used.

I'm not quite happy yet with some of the casts and the
suppression of unchecked warnings, so this is definitely a work in
progress.