Made the annotation values for @Node, @Attribute and @Source optional.
In Java 8, the names of parameters can be accessed through
reflection, provided that the code was compiled with the
"-parameters" option enabled.

This is a perfect candidate for code conciseness, or convention over
configuration.

If you parameters have the same name as the XML elements they
correspond to, then the annotation's values can be left out. Clean code!