Bugfix for parsers that return null for node content.
If a parser parses a String into a null value - which is perfectly
fine - Sprox would throw a NullPointerException in case of node content.

In case of an attribute, parsing into a null value was already working.

Added test cases for both nodes and attributes, and fixed the code for
nodes.